adding author profile
not sure how the icons work.  couldn't find an example of the fa-github-square as an image anywhere in the repo.  Looking at the running page, it seems to be rendered inline.

I can get icons for blogger and twitter, but where can I find the correct size / format and where to put the images?